requirements.txt
requests~=2.28.1
tqdm~=4.64.1
lxml~=4.9.1
Scrapy~=2.8.0
itemadapter~=0.7.0
pydantic~=1.10.2
beautifulsoup4~=4.11.1
PyYAML~=6.0